decorador funcionando sin ser generico

+++ Item.h
@@ -50,7 +50,7 @@ class Item { // Clase abstracta // Interfaz
 
         virtual void setPotencia(int potencia) = 0; // Metodo virtual puro
 
-        virtual double costo() = 0; // Metodo virtual puro
+        virtual int costo() = 0; // Metodo virtual puro
 
         virtual string toString() = 0; // Metodo virtual puro;
