+++ Llantas.h
@@ -36,7 +36,7 @@ public:
 
     void setPotencia(int potencia); // Metodo virtual puro
 
-    double costo();
+    int costo();
 
     string toString();
